Created reusable ScanSummary component, added thousand separator and removed decimal places for numeric data
